Added buttons to download the chart images
Chart images are downloaded via image URIs, which are stored in a variable when the charts are rendered.

Chart colors depend on the night mode setting. The chart is then set to have a transparent background for rendering on the webpage.

This update also brings linking to the formula modal. q=learn is the URL parameter to open the learnModal on load. An input to quickly copy this URL was added to the modal.

Lastly, added a link to a relevant 3Blue1Brown video.